fix: restore long NetworkRecoveryInterval binary descriptors (TRGN-4200)
Keep getNetworkRecoveryInterval()/setNetworkRecoveryInterval(long) as ()J/(J)V for 2.11.7 linkage. Use Long field plus OrNull/Long-setter for omitted-value binding validation; allow 0.

+++ sdk/trade360-java-sdk/src/main/java/eu/lsports/trade360_java_sdk/feed/rabbitmq/configurations/DynamicRabbitMQDefinitionRegistrar.java
@@ -80,7 +80,7 @@ public class DynamicRabbitMQDefinitionRegistrar implements BeanDefinitionRegistr
             factory.setConcurrentConsumers(rabbitConnectionConfiguration.getConcurrentConsumers());
             factory.setMaxConcurrentConsumers(rabbitConnectionConfiguration.getMaxConcurrentConsumers());
             factory.setPrefetchCount(rabbitConnectionConfiguration.getPrefetchCount());
-            factory.setRecoveryInterval(rabbitConnectionConfiguration.getNetworkRecoveryInterval().longValue());
+            factory.setRecoveryInterval(rabbitConnectionConfiguration.getNetworkRecoveryInterval());
             factory.setMissingQueuesFatal(false);
             return factory;
         });
